Move the AI skills page to /:countryId/ai-agents with a permanent redirect
The skills catalog serves both Claude Code and Codex, so the landing URL
drops the Claude branding: /:countryId/claude-plugin permanently redirects
to /:countryId/ai-agents. Hero eyebrow and page title go runtime-agnostic,
the install terminal gains the Codex path, and the sitemap entry moves.
